+++ app/Scripts/MejoraContinua/HD-3180/script/1_OpcionesUsuario_MC_HD-3180.sql
@@ -3,7 +3,9 @@ GO
 
 IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'OpcionesUsuario' AND type = 'U')
 begin
-	DROP TABLE [dbo].[UsuarioOpciones]
+	IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'UsuarioOpciones' AND type = 'U') 	
+		DROP TABLE [dbo].[UsuarioOpciones]
+
 	DROP TABLE [dbo].[OpcionesUsuario]
 end
 GO
@@ -23,7 +25,9 @@ GO
 
 IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'OpcionesUsuario' AND type = 'U')
 begin
-	DROP TABLE [dbo].[UsuarioOpciones]
+	IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'UsuarioOpciones' AND type = 'U') 	
+		DROP TABLE [dbo].[UsuarioOpciones]
+
 	DROP TABLE [dbo].[OpcionesUsuario]
 end
 GO
@@ -43,7 +47,9 @@ GO
 
 IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'OpcionesUsuario' AND type = 'U')
 begin
-	DROP TABLE [dbo].[UsuarioOpciones]
+	IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'UsuarioOpciones' AND type = 'U') 	
+		DROP TABLE [dbo].[UsuarioOpciones]
+
 	DROP TABLE [dbo].[OpcionesUsuario]
 end
 GO
@@ -63,7 +69,9 @@ GO
 
 IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'OpcionesUsuario' AND type = 'U')
 begin
-	DROP TABLE [dbo].[UsuarioOpciones]
+	IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'UsuarioOpciones' AND type = 'U') 	
+		DROP TABLE [dbo].[UsuarioOpciones]
+
 	DROP TABLE [dbo].[OpcionesUsuario]
 end
 GO
@@ -83,7 +91,9 @@ GO
 
 IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'OpcionesUsuario' AND type = 'U')
 begin
-	DROP TABLE [dbo].[UsuarioOpciones]
+	IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'UsuarioOpciones' AND type = 'U') 	
+		DROP TABLE [dbo].[UsuarioOpciones]
+
 	DROP TABLE [dbo].[OpcionesUsuario]
 end
 GO
@@ -103,7 +113,9 @@ GO
 
 IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'OpcionesUsuario' AND type = 'U')
 begin
-	DROP TABLE [dbo].[UsuarioOpciones]
+	IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'UsuarioOpciones' AND type = 'U') 	
+		DROP TABLE [dbo].[UsuarioOpciones]
+
 	DROP TABLE [dbo].[OpcionesUsuario]
 end
 GO
@@ -123,7 +135,9 @@ GO
 
 IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'OpcionesUsuario' AND type = 'U')
 begin
-	DROP TABLE [dbo].[UsuarioOpciones]
+	IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'UsuarioOpciones' AND type = 'U') 	
+		DROP TABLE [dbo].[UsuarioOpciones]
+
 	DROP TABLE [dbo].[OpcionesUsuario]
 end
 GO
@@ -143,7 +157,9 @@ GO
 
 IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'OpcionesUsuario' AND type = 'U')
 begin
-	DROP TABLE [dbo].[UsuarioOpciones]
+	IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'UsuarioOpciones' AND type = 'U') 	
+		DROP TABLE [dbo].[UsuarioOpciones]
+
 	DROP TABLE [dbo].[OpcionesUsuario]
 end
 GO
@@ -163,7 +179,9 @@ GO
 
 IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'OpcionesUsuario' AND type = 'U')
 begin
-	DROP TABLE [dbo].[UsuarioOpciones]
+	IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'UsuarioOpciones' AND type = 'U') 	
+		DROP TABLE [dbo].[UsuarioOpciones]
+
 	DROP TABLE [dbo].[OpcionesUsuario]
 end
 GO
@@ -183,7 +201,9 @@ GO
 
 IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'OpcionesUsuario' AND type = 'U')
 begin
-	DROP TABLE [dbo].[UsuarioOpciones]
+	IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'UsuarioOpciones' AND type = 'U') 	
+		DROP TABLE [dbo].[UsuarioOpciones]
+
 	DROP TABLE [dbo].[OpcionesUsuario]
 end
 GO
@@ -203,7 +223,9 @@ GO
 
 IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'OpcionesUsuario' AND type = 'U')
 begin
-	DROP TABLE [dbo].[UsuarioOpciones]
+	IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'UsuarioOpciones' AND type = 'U') 	
+		DROP TABLE [dbo].[UsuarioOpciones]
+
 	DROP TABLE [dbo].[OpcionesUsuario]
 end
 GO
@@ -223,7 +245,9 @@ GO
 
 IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'OpcionesUsuario' AND type = 'U')
 begin
-	DROP TABLE [dbo].[UsuarioOpciones]
+	IF EXISTS (SELECT 1 FROM sysobjects WHERE name = 'UsuarioOpciones' AND type = 'U') 	
+		DROP TABLE [dbo].[UsuarioOpciones]
+
 	DROP TABLE [dbo].[OpcionesUsuario]
 end
 GO
